[feature #165020128] put the checkSlug function in helpers, add commentsCount in the payload of getting all comments

+++ server/controllers/comments.js
@@ -1,34 +1,12 @@
-import { Comment, Article, User } from '../models';
+import { Comment, User } from '../models';
 import errorHandler from '../helpers/errorHandler';
+import checkSlug from '../helpers/checkSlug';
 /**
  *
  *
  * @class Comments
  */
 class Comments {
-  /**
-     *
-     * returning the article giving the slug.
-     * @static
-     * @param {object} req
-     * @param {object} res
-     * @returns {id} article id
-     * @memberof Comments
-     */
-  static async checkSlug(req, res) {
-    const articleId = await Article.findOne({
-      where: { slug: req.params.slug },
-      attributes: ['id']
-    });
-    if (!articleId) {
-      return res.status(404).json({
-        status: 404,
-        message: 'Article is not found.'
-      });
-    }
-    return articleId.id;
-  }
-
   /**
      *
      *
@@ -40,8 +18,10 @@ class Comments {
      */
   static async create(req, res) {
     try {
-      const articleId = await Comments.checkSlug(req, res);
-
+      const articleId = await checkSlug(req, res);
+      if (typeof articleId !== 'number') {
+        return false;
+      }
       const userId = req.user.id;
       const comment = {
         body: req.body.comment.body,
@@ -71,9 +51,11 @@ class Comments {
      */
   static async getAll(req, res) {
     try {
-      const articleId = await Comments.checkSlug(req, res);
-
-      const result = await Comment.findAll({
+      const articleId = await checkSlug(req, res);
+      if (typeof articleId !== 'number') {
+        return false;
+      }
+      const result = await Comment.findAndCountAll({
         where: { articleId, },
         attributes: ['id', 'body', 'createdAt', 'updatedAt'],
         include: [{
@@ -84,7 +66,10 @@ class Comments {
       });
       return res.status(200).json({
         status: 200,
-        data: { comments: result }
+        data: {
+          comments: result.rows,
+          commentsCount: result.count,
+        }
       });
     } catch (e) {
       errorHandler.errorResponse(res, e);
@@ -102,7 +87,10 @@ class Comments {
      */
   static async delete(req, res) {
     try {
-      await Comments.checkSlug(req, res);
+      const articleId = await checkSlug(req, res);
+      if (typeof articleId !== 'number') {
+        return false;
+      }
       const { id } = req.params;
       const result = await Comment.destroy({ where: { id, }, returning: true });
 
@@ -133,7 +121,10 @@ class Comments {
      */
   static async update(req, res) {
     try {
-      await Comments.checkSlug(req, res);
+      const articleId = await checkSlug(req, res);
+      if (typeof articleId !== 'number') {
+        return false;
+      }
       const { id } = req.params;
       const { body } = req.body.comment;
       const result = await Comment.update({ body, }, {
